+++ CHANGELOG.rst
@@ -2,9 +2,11 @@ Changelog
 =========
 
 10.0.0, 2024-12-20
------------------
+------------------
 - Added ``miscellany.restrict_to_trips`` and used it as a helper function to simplify the other restriction functions.
 - Breaking change: Removed the validation module ``validators.py`` to avoid duplicating the work of what is now `the canonical feed validator <https://github.com/MobilityData/gtfs-validator>`_ (written in Java).
+- Breaking change: Changed ``feed.write`` to ``feed.to_file`` and stopped default rounding.
+- Breaking change: Changed ``miscellany.summarize`` to ``feed.list_fields`` and stopped default rounding.
 
 9.0.0, 2024-12-19
 -----------------